move synthetics status update function to synthetics.go from utils.go (and associated test)
